Move Cost / Constraint creation functionality outside of MathematicalProgram.
