+++ drake/solvers/CMakeLists.txt
@@ -30,6 +30,8 @@ list(APPEND optimization_files
   ${solver_headers}
   constraint.cc
   cost.cc
+  create_constraint.cc
+  create_cost.cc
   decision_variable.cc
   equality_constrained_qp_solver.cc
   linear_system_solver.cc
@@ -80,6 +82,7 @@ drake_install_headers(
   binding.h
   constraint.h
   cost.h
+  create_constraint.h
   create_cost.h
   function.h
   mathematical_program.h
