+++ drake/solvers/mathematical_program.h
@@ -28,6 +28,7 @@
 #include "drake/solvers/binding.h"
 #include "drake/solvers/constraint.h"
 #include "drake/solvers/cost.h"
+#include "drake/solvers/create_constraint.h"
 #include "drake/solvers/create_cost.h"
 #include "drake/solvers/decision_variable.h"
 #include "drake/solvers/function.h"
@@ -660,7 +661,7 @@ class MathematicalProgram {
    */
   template <typename F>
   static std::shared_ptr<Cost> MakeCost(F&& f) {
-    return CreateFunctionCost(f);
+    return MakeFunctionCost(f);
   }
 
   /**
@@ -685,7 +686,7 @@ class MathematicalProgram {
   typename std::enable_if<detail::is_cost_functor_candidate<F>::value,
                           Binding<Cost>>::type
   AddCost(F&& f, const Eigen::Ref<const VectorXDecisionVariable>& vars) {
-    auto c = CreateFunctionCost(std::forward<F>(f));
+    auto c = MakeFunctionCost(std::forward<F>(f));
     return AddCost(c, vars);
   }
 
@@ -827,8 +828,7 @@ class MathematicalProgram {
       const Eigen::Ref<const Eigen::MatrixXd>& A,
       const Eigen::Ref<const Eigen::VectorXd>& b,
       const Eigen::Ref<const VectorXDecisionVariable>& vars) {
-    return AddQuadraticCost(2 * A.transpose() * A, -2 * A.transpose() * b,
-                            vars);
+    return AddCost(MakeL2NormCost(A, b), vars);
   }
 
   /**
@@ -1055,54 +1055,10 @@ class MathematicalProgram {
    */
   template <typename Derived>
   typename std::enable_if<
-      std::is_base_of<Eigen::ArrayBase<Derived>, Derived>::value &&
-          std::is_same<typename Derived::Scalar, symbolic::Formula>::value,
+      detail::is_eigen_scalar_same<Derived, symbolic::Formula>::value,
       Binding<LinearConstraint>>::type
-  AddLinearConstraint(const Derived& formulas) {
-    const auto n = formulas.rows() * formulas.cols();
-
-    // Decomposes 2D-array of formulas into 1D-vector of expression, `v`, and
-    // two 1D-vector of double `lb` and `ub`.
-    constexpr int flat_vector_size{
-        MultiplyEigenSizes<Derived::RowsAtCompileTime,
-                           Derived::ColsAtCompileTime>::value};
-    Eigen::Matrix<symbolic::Expression, flat_vector_size, 1> v{n};
-    Eigen::Matrix<double, flat_vector_size, 1> lb{n};
-    Eigen::Matrix<double, flat_vector_size, 1> ub{n};
-    int k{0};  // index variable for 1D components.
-    for (int i{0}; i < formulas.rows(); ++i) {
-      for (int j{0}; j < formulas.cols(); ++j, ++k) {
-        const symbolic::Formula& f{formulas(i, j)};
-        if (is_equal_to(f)) {
-          // f(i) := (lhs == rhs)
-          //         (lhs - rhs == 0)
-          v(k) = get_lhs_expression(f) - get_rhs_expression(f);
-          lb(k) = 0.0;
-          ub(k) = 0.0;
-        } else if (is_less_than_or_equal_to(f)) {
-          // f(i) := (lhs <= rhs)
-          //         (-∞ <= lhs - rhs <= 0)
-          v(k) = get_lhs_expression(f) - get_rhs_expression(f);
-          lb(k) = -std::numeric_limits<double>::infinity();
-          ub(k) = 0.0;
-        } else if (is_greater_than_or_equal_to(f)) {
-          // f(i) := (lhs >= rhs)
-          //         (∞ >= lhs - rhs >= 0)
-          v(k) = get_lhs_expression(f) - get_rhs_expression(f);
-          lb(k) = 0.0;
-          ub(k) = std::numeric_limits<double>::infinity();
-        } else {
-          std::ostringstream oss;
-          oss << "MathematicalProgram::AddLinearConstraint is called with an "
-                 "array of formulas which includes a formula "
-              << f
-              << " which is not a relational formula using one of {==, <=, >=} "
-                 "operators.";
-          throw std::runtime_error(oss.str());
-        }
-      }
-    }
-    return AddLinearConstraint(v, lb, ub);
+  AddLinearConstraint(const Eigen::ArrayBase<Derived>& formulas) {
+    return AddConstraint(internal::ParseLinearConstraint(formulas));
   }
 
   /**
@@ -1172,17 +1128,11 @@ class MathematicalProgram {
    */
   template <typename DerivedV, typename DerivedB>
   typename std::enable_if<
-      std::is_base_of<Eigen::MatrixBase<DerivedV>, DerivedV>::value &&
-          std::is_base_of<Eigen::MatrixBase<DerivedB>, DerivedB>::value &&
-          std::is_same<typename DerivedV::Scalar,
-                       symbolic::Expression>::value &&
-          std::is_same<typename DerivedB::Scalar, double>::value &&
-          (DerivedV::ColsAtCompileTime == 1 ||
-           DerivedB::ColsAtCompileTime == 1),
+      detail::is_eigen_vector_expression_double_pair<DerivedV, DerivedB>::value,
       Binding<LinearEqualityConstraint>>::type
   AddLinearEqualityConstraint(const Eigen::MatrixBase<DerivedV>& v,
                               const Eigen::MatrixBase<DerivedB>& b) {
-    return DoAddLinearEqualityConstraint(v, b);
+    return AddConstraint(internal::ParseLinearEqualityConstraint(v, b));
   }
 
   /**
@@ -1206,65 +1156,13 @@ class MathematicalProgram {
    */
   template <typename DerivedV, typename DerivedB>
   typename std::enable_if<
-      std::is_base_of<Eigen::MatrixBase<DerivedV>, DerivedV>::value &&
-          std::is_base_of<Eigen::MatrixBase<DerivedB>, DerivedB>::value &&
-          std::is_same<typename DerivedV::Scalar,
-                       symbolic::Expression>::value &&
-          std::is_same<typename DerivedB::Scalar, double>::value &&
-          DerivedV::ColsAtCompileTime != 1 && DerivedB::ColsAtCompileTime != 1,
+      detail::is_eigen_matrix_expression_double_pair<DerivedV, DerivedB>::value,
       Binding<LinearEqualityConstraint>>::type
   AddLinearEqualityConstraint(const Eigen::MatrixBase<DerivedV>& V,
                               const Eigen::MatrixBase<DerivedB>& B,
                               bool lower_triangle = false) {
-    if (lower_triangle) {
-      DRAKE_DEMAND(V.rows() == V.cols() && B.rows() == B.cols());
-    }
-    DRAKE_DEMAND(V.rows() == B.rows() && V.cols() == B.cols());
-
-    // Form the flatten version of V and B, when lower_triangle = false,
-    // the flatten version is just to concatenate each column of the matrix;
-    // otherwise the flatten version is to concatenate each column of the
-    // lower triangular part of the matrix.
-    const int V_rows = DerivedV::RowsAtCompileTime != Eigen::Dynamic
-                           ? static_cast<int>(DerivedV::RowsAtCompileTime)
-                           : static_cast<int>(DerivedB::RowsAtCompileTime);
-    const int V_cols = DerivedV::ColsAtCompileTime != Eigen::Dynamic
-                           ? static_cast<int>(DerivedV::ColsAtCompileTime)
-                           : static_cast<int>(DerivedB::ColsAtCompileTime);
-
-    if (lower_triangle) {
-      const int V_triangular_size =
-          V_rows != Eigen::Dynamic ? (V_rows + 1) * V_rows / 2 : Eigen::Dynamic;
-      int V_triangular_size_dynamic = V.rows() * (V.rows() + 1) / 2;
-      Eigen::Matrix<symbolic::Expression, V_triangular_size, 1> flat_lower_V(
-          V_triangular_size_dynamic);
-      Eigen::Matrix<double, V_triangular_size, 1> flat_lower_B(
-          V_triangular_size_dynamic);
-      int V_idx = 0;
-      for (int j = 0; j < V.cols(); ++j) {
-        for (int i = j; i < V.rows(); ++i) {
-          flat_lower_V(V_idx) = V(i, j);
-          flat_lower_B(V_idx) = B(i, j);
-          ++V_idx;
-        }
-      }
-      return AddLinearEqualityConstraint(flat_lower_V, flat_lower_B);
-    } else {
-      const int V_size = V_rows != Eigen::Dynamic && V_cols != Eigen::Dynamic
-                             ? V_rows * V_cols
-                             : Eigen::Dynamic;
-      Eigen::Matrix<symbolic::Expression, V_size, 1> flat_V(V.size());
-      Eigen::Matrix<double, V_size, 1> flat_B(V.size());
-      int V_idx = 0;
-      for (int j = 0; j < V.cols(); ++j) {
-        for (int i = 0; i < V.rows(); ++i) {
-          flat_V(V_idx) = V(i, j);
-          flat_B(V_idx) = B(i, j);
-          ++V_idx;
-        }
-      }
-      return AddLinearEqualityConstraint(flat_V, flat_B);
-    }
+    return AddConstraint(
+        internal::ParseLinearEqualityConstraint(V, B, lower_triangle));
   }
 
   /** AddLinearEqualityConstraint
@@ -1332,8 +1230,8 @@ class MathematicalProgram {
   Binding<LinearEqualityConstraint> AddLinearEqualityConstraint(
       const Eigen::Ref<const Eigen::RowVectorXd>& a, double beq,
       const VariableRefList& vars) {
-    return AddLinearEqualityConstraint(a, beq,
-                                       ConcatenateVariableRefList(vars));
+    return AddConstraint(std::make_shared<LinearEqualityConstraint>(a, beq),
+                         ConcatenateVariableRefList(vars));
   }
 
   /**
@@ -2487,10 +2385,6 @@ class MathematicalProgram {
       }
     }
   }
-
-  Binding<LinearEqualityConstraint> DoAddLinearEqualityConstraint(
-      const Eigen::Ref<const VectorX<symbolic::Expression>>& v,
-      const Eigen::Ref<const Eigen::VectorXd>& b);
 
   // Adds a linear constraint represented by a set of symbolic formulas to the
   // program.
