+++ drake/systems/trajectory_optimization/direct_trajectory_optimization.h
@@ -133,7 +133,7 @@ class DirectTrajectoryOptimization : public solvers::MathematicalProgram {
   typename std::enable_if<solvers::detail::is_cost_functor_candidate<F>::value,
                           std::shared_ptr<solvers::Cost>>::type
   AddRunningCostFunc(F&& f) {
-    auto c = solvers::CreateFunctionCost(std::forward<F>(f));
+    auto c = solvers::MakeFunctionCost(std::forward<F>(f));
     AddRunningCost(c);
     return c;
   }
@@ -268,7 +268,7 @@ class DirectTrajectoryOptimization : public solvers::MathematicalProgram {
   typename std::enable_if<solvers::detail::is_cost_functor_candidate<F>::value,
                           std::shared_ptr<solvers::Cost>>::type
   AddFinalCostFunc(F&& f) {
-    auto c = solvers::CreateFunctionCost(std::forward<F>(f));
+    auto c = solvers::MakeFunctionCost(std::forward<F>(f));
     AddFinalCost(c);
     return c;
   }
